+++ plugins/kv/config.go
@@ -1,5 +1,6 @@
 package kv
 
+// Config represents general storage configuration with keys as the user defined kv-names and values as the drivers
 type Config struct {
 	Data map[string]interface{} `mapstructure:"kv"`
 }