+++ plugins/kv/drivers/boltdb/plugin.go
@@ -63,3 +63,6 @@ func (s *Plugin) Provide(key string) (kv.Storage, error) {
 func (s *Plugin) Name() string {
 	return PluginName
 }
+
+// Available interface implementation
+func (s *Plugin) Available() {}