+++ plugins/kv/drivers/memory/plugin.go
@@ -62,3 +62,7 @@ func (s *Plugin) Provide(key string) (kv.Storage, error) {
 func (s *Plugin) Name() string {
 	return PluginName
 }
+
+// Available interface implementation
+func (s *Plugin) Available() {
+}