+++ plugins/kv/storage.go
@@ -184,3 +184,7 @@ func (p *Plugin) RPC() interface{} {
 func (p *Plugin) Name() string {
 	return PluginName
 }
+
+// Available interface implementation
+func (p *Plugin) Available() {
+}