+++ plugins/reload/plugin.go
@@ -161,3 +161,7 @@ func (s *Plugin) Stop() error {
 func (s *Plugin) Name() string {
 	return PluginName
 }
+
+// Available interface implementation
+func (s *Plugin) Available() {
+}