+++ plugins/resetter/plugin.go
@@ -74,6 +74,10 @@ func (p *Plugin) Name() string {
 	return PluginName
 }
 
+// Available interface implementation
+func (p *Plugin) Available() {
+}
+
 // RPC returns associated rpc service.
 func (p *Plugin) RPC() interface{} {
 	return &rpc{srv: p, log: p.log}