+++ plugins/rpc/plugin.go
@@ -122,6 +122,10 @@ func (s *Plugin) Name() string {
 	return PluginName
 }
 
+// Available interface implementation
+func (s *Plugin) Available() {
+}
+
 // Collects all plugins which implement Name + RPCer interfaces
 func (s *Plugin) Collects() []interface{} {
 	return []interface{}{